Refactors vertical line
